feat: opt-out ssm parameters for github app (#4335)
Hi! 👋

Throughout my recent experiences, we had strong requirements to **not**
have secret values in the state. To handle that, I suggest to make
optional the creation of SSM parameters as part of the regular flow so
it doesn't leak secrets in Terraform state.

We could alternatively use the latest
[`ephemeral`](https://developer.hashicorp.com/terraform/language/resources/ephemeral)
feature, but I'm not sure everyone is using Terraform 1.10+ atm since
it's quite recent. I also think relying on ephemerals should be part of
a new breaking release as it won't be compatible with Terraform < 1.10.

Let me know what you think of this!

---------

Co-authored-by: Niek Palm <npalm@users.noreply.github.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Niek Palm <niek.palm@philips.com>