fix: Safer chown call in runner scripts for custom AMIs (#4654)
Updated the install-runner and start-runner scripts to call `chown` on
explicit path instead of just using the current directory.

The current script will work for most users, however, this change will
make it more safer and avoid risk of unintended changes. For example, if
the `cd /opt/actions-runner` call in the template file
(`images/start-runner.sh`) is missed, the script will change owner of
the root directory (`/`) instead.

---------

Co-authored-by: Niek Palm <npalm@users.noreply.github.com>